+++ src/Monolog/Service/ComponentFactory.php
@@ -4,6 +4,8 @@ declare(strict_types = 1);
 
 namespace OctoLab\Common\Monolog\Service;
 
+use OctoLab\Common\Util\ArrayHelper;
+
 /**
  * @author Kamil Samigullin <kamil@samigullin.info>
  */
@@ -15,38 +17,38 @@ class ComponentFactory
     public static function withDefaults(): ComponentFactory
     {
         return new self(
-            (new Component())
+            (new ComponentBuilder())
                 ->setClass('Monolog\Logger')
                 ->addDependency('handlers', 'pushHandler')
                 ->addDependency('processors', 'pushProcessor'),
-            (new Component())
+            (new ComponentBuilder())
                 ->setNamespace('Monolog\Handler')
                 ->setSuffix('Handler')
                 ->addDependency('processors', 'pushProcessor')
                 ->addDependency('formatter', 'setFormatter'),
-            (new Component())
+            (new ComponentBuilder())
                 ->setNamespace('Monolog\Formatter')
                 ->setSuffix('Formatter'),
-            (new Component())
+            (new ComponentBuilder())
                 ->setNamespace('Monolog\Processor')
                 ->setSuffix('Processor')
         );
     }
 
-    /** @var Component[] */
+    /** @var ComponentBuilder[] */
     private $components;
 
     /**
-     * @param Component $channel
-     * @param Component $handler
-     * @param Component $formatter
-     * @param Component $processor
+     * @param ComponentBuilder $channel
+     * @param ComponentBuilder $handler
+     * @param ComponentBuilder $formatter
+     * @param ComponentBuilder $processor
      */
     public function __construct(
-        Component $channel,
-        Component $handler,
-        Component $formatter,
-        Component $processor
+        ComponentBuilder $channel,
+        ComponentBuilder $handler,
+        ComponentBuilder $formatter,
+        ComponentBuilder $processor
     ) {
         $this->components = [
             'channels' => $channel,
@@ -54,5 +56,44 @@ class ComponentFactory
             'formatters' => $formatter,
             'processors' => $processor,
         ];
+    }
+
+    /**
+     * @param array $config
+     *
+     * @return mixed
+     */
+    public function build(array $config)
+    {
+        $default = [
+            'class' => null,
+            'type' => null,
+            'arguments' => [],
+            '_key' => null,
+        ];
+        $config = ArrayHelper::merge($default, $config);
+    }
+
+    /**
+     * @param string $key
+     *
+     * @return string[]
+     *
+     * @throws \InvalidArgumentException
+     */
+    public function getDependencies(string $key): array
+    {
+        if (!isset($this->components[$key])) {
+            throw new \InvalidArgumentException(sprintf('No component with key "%s" found.', $key));
+        }
+        return $this->components[$key]->getDependencies();
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getAvailableComponentKeys(): array
+    {
+        return array_keys($this->components);
     }
 }